refactor: convert utilities/randos* to typescript
