+++ src/layer/random.js
@@ -1,5 +1,5 @@
 const { Model } = require('./types');
-const randos2D = require('../utilities/randos-2d');
+const { randos2D } = require('../utilities/randos');
 const { zeros2D } = require('../utilities/zeros-2d');
 
 class Random extends Model {
